refactor(syncstore): the write side is named apart from the read side (#20)
CORE-17. `zonestore` and `zonesyncstore` differed by four letters in the
middle of a word, on the one boundary this design calls load-bearing: a
handler may import the read side and may never reach the write side, and
the whole question a reader has on the import line is which of the two
they are looking at. A name that must be read to its end to be told from
its opposite is a name that gets misread.

`internal/zonesyncstore` becomes `internal/syncstore`, which differs from
`internal/zonestore` at its first character. Only the write side moves:
the read side is the name a handler reaches for, and renaming both would
double the surface to settle one confusion.

The AC2 import invariant and its exemption list follow the rename —
`offrequestpath_test.go`'s syncAdapter names the new path, so the check
still refuses rather than passes when the adapter is missing, and the
exemption list is still the same two packages it was.
